make RapidJSON a private property for DOTS target
It generally does not make sense to have libraries as public properties
for a target if they are provided by Git sub-modules. Since those
dependencies are not necessarily installed on the system, linking
against the target might fail, because the dependencies cannot be
found.

To make a dependency from a Git sub-module public, it has to be
installed to the system alongside the main target itself.