move code generation to dedicated project
The different components and dependencies for the type generation from
model files is currently scattered across multiple files. The model
sub-project only contains the templates, but does not provide any means
of actually using the code generator, which itself is provided by a Git
sub-module. Additionally, the CMake functions to actually use the code
generator are located in both the FindDotsCodeGenerator, as well as the
previously removed FindDOTS modules.

The structure and locality of the code generation can be improved by
moving all necessary components into a dedicated modern CMake project.

The solution applied in this commit has several benefits. Not only can
the project be used like any other config-based CMake project, it also
bundles all dependencies in a single location and defines the CMake
code generation functions only once.