remove out of initializer move in prop construct
Moving out of the initializer rvalue should be legal but can generate a
segmentation fault when compiling with GCC 8 and the O2 and LTO
optimization flags. It is currently unclear if this is indeed
permitted by some special rule or whether the compiler performs an
illegal optimization in this particular situation.

Since this issue does not occur with GCC 7.3 it might be desirable to
perform additional investigations with GCC 9 and or Clang at some point
in the future.