rename ASIO single shot timer for clarity
A single shot timer is per definition a timer that only runs once. It
is therefore unintuitive that it would have a periodic option and thus
entirely nullify the concept of a single shot timer.

Instead, it makes more sense to have a general timer that can be
single shot or not depending on what the user chooses.