replace boost ASIO with local standalone version
Since ASIO was heavily improved over the last year, it is highly
desirable to update it to a recent version.

Using the ASIO version from boost has some severe drawbacks. Since the
CMake support of boost is bound to the CMake version, it is not easily
possible to increase the ASIO version without also requiring a fairly
high version of CMake. It is also not possible to update boost for
non-ASIO reasons without also updating ASIO and vice versa.

Additionally, the standalone ASIO version can
easily be imported as a sub-module, thus laying the groundwork of
making the DOTS library independent of any libraries that have to be
installed on the system in a specific version.

Finally, since ASIO is on its way of becoming a part of the standard,
it would be much easier to make the switch from a self-contained module
than from an external library.