use ASIO for Qt applications
The current solution to use DOTS with Qt applications requires a lot of
code duplication for various IO components. This creates an unnecessary
potential for error, which might lead to Qt applications behave
differently, because their IO processing is not the same as that of
non-Qt applications. Additionally, Qt applications lose all the
efficiency and features ASIO has to offer.

The approach also makes DOTS dependent on the Qt library and an
application that wants to use Qt with DOTS has to use a special version
of the DOTS library where Qt support was enabled during build time.

This commit completely removes the Qt dependency from the library and
instead adds support to run the DOTS application and event loop in a
dedicated Qt timer with a timeout of zero. These kind of timers are
treated in a special way in Qt and are guaranteed to always be executed
after all GUI events have been processed (see documentation of QTimer).

This way, a Qt application can run the regular DOTS ASIO event loop in
a single-threaded asynchronous way and at a well-defined point in time,
without having to sacrifice any responsiveness in the GUI. The approach
should also allow for easy integration of DOTS into other event loop
frameworks.