move event loop components into ASIO services
The current IO component architecture has several flaws. It relies
heavily on the state of multiple global objects, which have to be
initialized in the right order to function properly. This makes the
control flow very unintuitive and hard to maintain.

Using ASIO services is a first step for a clearer structure of IO the
components. Every IO functionality is managed by a dedicated service,
which is tied to the global IO context. Because services can be
optional and are interchangeable, they provide a flexible way of
changing IO functionality without sacrificing a clear control flow.