make listening a functionality of the IO services
Listening for TCP connections is currently only a functionality of the
DOTS daemon. By making it a functionality of the IO services in the
DOTS library with an additional abstraction layer, the daemon is now
ready to be made agnostic about the underlying communication, working
entirely on a listener interface.