rename DOTS socket to channel
To maintain a certain amount of flexibility, the type of data
communication should be completely abstracted from the DOTS components.

Because there are a great amount of potential communication types that
could be added in the future, the name of the abstraction layer should
reflect something generic that is not already associated with certain
communication technologies.