remove remaining TCP exposition from channel
The channel currently still exposes TCP specific dependencies by
requiring the sub-class to implement a connect function with hostname
and port parameters, which alternative channel types might conceptually
be unaware of.

This commit changes the channel to a completely communication type
agnostic interface, requiring the sub-class to only implement general
asynchronous read and transmit functions. The concept of connecting and
disconnecting also has been removed and should instead be handled via
RAII in the implementing classes.