introduce communication transmissions
By requiring the handling of a payload, all communication channels are
currently still bound to some form of direct data serialization. To
further increase the means of communication and serialization
abstraction, the channel now works on the basis of transmissions.

Transmissions encapsulate a DOTS instance for communication with a
channel. It connects the actual instance that is transported with an id
that is unique for the local DOTS application. Since communication
channels can use a variety of serialization methods or no serialization
at all, this id can be used to identify transmission and cache
serialized data if necessary, so it does not have to be serialized
multiple times.

This commit also removes the remaining CBOR dependency in various
components.