introduce virtual channel
There are a lot of situations where communication with an actual DOTS
daemon is not desired or reasonable. A few examples would be
non-server-based ways of communication (e.g. files) or testing
environments (e.g. mocking).

This commit introduces a virtual channel, that is capable of mimicking
a server connection, including the DOTS handshake. The channel can be
used both to eavesdrop on the communication, but also to spoof
transmissions, which for the client are indistinguishable from the ones
coming from an actual DOTS daemon.

The class is supposed to be usable on its own, as well as as a base
class for additional virtual channel types.