safe-guard multiple async receive calls on channel
Multiple calls to async receive on a single channel do not make logical
sense. In particular, ASIO does not allow multiple async operations on
a single resource.

It would be possible to allow multiple handlers to process the same
async receive result, however, since that is currently not a use case,
async receive calls should be limited to one at the same time.