add default value for TCP listener backlog option
The currently used value of the backlog option for the TCP listener was
chosen more or less arbitrarily.

Unless there are reasons to explicitly restrict the number of listen
connections, this value should be as high as possible on the current
system.

This commit allows the user to not specify a backlog parameter at all,
in which case the listener will use the default value defined by ASIO.