define two invalid properties as equal
Currently two properties are defined to be unequal if they are both
invalid. This is unintuitive, since an invalid property is conceptually
similar to a null pointer or an empty optional, which both compare
equal when compared to another null or empty instance. It also prevents
struct equality from being defined as equality of all properties, which
would be the natural approach.