add included properties support for equal and less
Currently equal and less comparisons do not allow the user to specify
included properties. This is an inconsistency with the other struct
operations, which support this option.

This commit adjusts the less and equal operations to allow the
specification of included properties, which default to all properties
in the instance. It also fixes an inconsistency with the struct base
template, in which the less operation is only applying to the key
properties. Instead, the less operation is now consistent with the
struct base class and the less comparison on keys is restricted to the
less operator.