overhaul dispatcher and containers
The current implementation of the DOTS containers is convoluted and
redundant. Since explicitly typed and typeless containers are
independent components, a lot of code is duplicated and yet not
equivalent at the same time. They also somewhat violate the single
responsibility principle, by functioning both as a data structure for
instances, but also as processors for received DOTS transmissions at
the same time. This leads, among other things, to the common problem of
having to fake received transmissions in certain situations, when no
actual receive took place. Another consequence is that certain
applications have to apply container processing manually, instead of
being able to use a central mechanism.

This commit completely overhauls the concept of containers, the
dispatcher and subscriptions as follows:

Containers are now pure data structures to store received instances,
along with meta information, such as creation and update times. All
containers, including explicitly typed ones, are now using the same
base implementation for inserting and removing instances. Instead of
having their own implementation, explicitly typed containers are now
just typed wrappers around the typeless base class.

All processing that was previously part of the container logic, is now
handled by the overhauled dispatcher, whose responsibility is the
handling of all subscriptions, along with the containers of the
subscribed types. This is a continuation of the effort to remove global
state in the library, since all globally accessible containers are now
part of the global dispatcher, which is owned by the global
transceiver.

Instead of having to manually add receivers, applications can now
exclusively use subscriptions, even for dynamic types. Since some
applications might not want the overhead of container processing,
subscriptions handlers can now be registered either for events or for
received transmissions.

Finally, subscription objects now use RAII to manage the subscription
state.

Note that this commit does not include the required changes for the
xdots and qdots utilities, which will be broken until fixed.