fix premature container access in xdots and qdots
The current design of xdots and qdots requires them to access empty
containers before they are created in the dispatcher. Since the
overhauled design does not allow non-const access to the container
pool, the containers do not yet exist and also cannot be created
prematurely.

This commit temporarily adds support for automatic container creation
to const instances of a container pool, so the utilities work again.

However, this should only be considered a temporary workaround, until
the utilities are properly overhauled.