adjust property set constant suffix for clarity
Currently the property set constants in each property are specified via
a _v suffix. This could be confusing, as _v suffixes are usually used
to refer to a value of a suffixed type. Since the property set is not a
value of the property type, it would be better to denote them by the
unique suffix _p, which does not have any other common meaning and is
also more consistent with the _i suffix of the property initializer
type.