add overhauled property set
Currently, the property set exposes operations that are typical for
both a mathematical set, as well as a bit set. For example, two sets
can be merged both with the + and the | operator. This leads to
inconsistent usage and decreased readability. Additionally, the
property set has an unintuitive < operator, which only compares the
stored bit set values.

Since the fact that it is implemented as bit set is an implementation
detail anyway, it would be better to only allow set semantics when
using a property set. This commit adds an inactive overhauled version
of the property set with semantics that mimic a mathematical set. The
class now exclusively supports the + and - operators to merge and
builds differences of sets and the < and > operators check for subset
properties.