add overhauled descriptor base class
The way descriptors are currently implemented has several fundamental
issues.

Because all descriptors are immediately added to the registry upon
construction, they are tightly coupled not only to the global registry
state, but also to the registry class itself, including all of its
dependencies. This makes it very difficulty to change and test
descriptors on their own. Additionally, the implementation does not
support move semantics and the focus on void pointers make its usage
cumbersome and error-prone.

This commit adds a new descriptor class with an overhauled concept. The
class is a template that must be specialized for every type that should
participate in the descriptor type system. The provided typeless
specialization hereby acts as a base class and requires sub-classes to
implement several core type operations, such as construction,
destruction and assignment.

Because every type has its own descriptor specialization, it is very
easy to access the descriptor of a type, just by the type alone.

The class is now also completely independent of the type registry and
instead of using void pointers, the interface now exclusively uses the
typeless class for safe and explicit handling of typeless references.