introduce static descriptor base class
The descriptors for fundamental (aka base) types are currently provided
by the StandardTypeDescriptor. This class mostly relays the core
descriptor operations to the explicit type via casting. Even though the
general approach is viable, it always completely erases all type
information. This makes their usage unattractive in situations where
the type is known and leads to always requiring checks and redundant
operation code for the static and dynamic case (see descriptor usage in
the Property class).

The current implementation is also outdated and, among other things,
does not support construction or assignment with parameters.

This commit introduces a base class template that provides a basic
descriptor implementation for an arbitrary static type. Contrary to the
StandardTypeDescriptor, it implements typed versions of the descriptor
operations and only relays the virtual overrides. This allows the
operations to be used efficiently without abstraction in cases where
the type is known, while only requiring one additional virtual
indirection in the typeless case.

Additionally, it can be used by non-fundamental types as well and also
centralizes access to one unique descriptor instance, for every static
type.