add overhauled DOTS vector class
The current DOTS vector class was built on the basis of a ripped STL
implementation, which was modified to provide a typeless base class.
This was necessary, because DOTS vectors need to be usable in typeless
contexts, while STL containers are inherently always explicitly typed.

This approach has a lot of downsides. Because it is based on a STL
implementation, it requires a lot of helper classes for allocators and
iterators, which also had to be re-implemented in crude way. Because of
this, changes are error-prone and it generally takes a lot of effort to
maintain anything.

This lead to the class becoming outdated and thus not benefiting from
potential improvements in a compiler specified version. It also means
less optimization potential for the compiler, which can often treat its
own types in a special way.

Finally, since it is not really compatible with std::vector, it
sometimes requires unnecessarily expensive conversions between
container types.

To fix these problems, this commit adds an overhauled version of the
DOTS vector class. Instead of re-implementing a complete vector, it
just acts as thin typeless wrapper around a native STL vector
implementation. This means that it is fully compatible to std::vector
and does not have to be updated manually, when the STL vector changes.
Each typed vector has a typeless base class, which relays a typeless
operations to the underlying typed implementation.

Note that this approach does no longer support the construction
of pure typeless vectors.