introduce overhauled fundamental types
The descriptor definitions for fundamental (aka base) types are
currently scattered around in different files. Since they are the core
building blocks for all other types, they should be defined in a
central place.

This commit adds a header for all fundamental types, which are defined
as aliases in the types namespace. These aliases can now be used to
explicitly refer to a specific type with the corresponding name from
the DOTS specification and without having to unintuitively use other
namespaces. This also makes it easier to exchange types in the future.

The header then defines static descriptors for all fundamental types.