add overhauled property base class
Because the valid property set is held by the struct base class, the
property base class is currently directly dependent on it. The struct
base in itself depends on the property iterator, which is further
dependent on the proxy property, which, by being a property, again
depends on the property base class.

Even though this circular dependency is mitigated by using forward
declarations, it is still inherently there and severely reduces the
maintainability and testability of properties in general.

This commit adds a new version of the property base class with an
overhauled approach to properties as a whole.

Firstly, it introduces the concept of a property area, which holds the
valid property set for a certain number of properties. It can either be
used as a base class or placed before property members in another type
and contains helper functions to access properties based on their
offset. The class only depends on the property set and therefore does
not require any knowledge of the struct base class and its dependencies
at all.

All the property data that is currently held by the old struct property
class and a property description data class, in the overhauled version
is replaced by a property descriptor, which also handles calculation
for the offset. It has both a typed and typeless version and allows
efficient access to the appropriate type descriptor in both scenarios.

Due to the changes mentioned above and the revised descriptors, the
overhauled property is much more simple and no longer contains
redundant code and additional checks for the typeless version. Instead,
all operations emerge from the corresponding property descriptor and
its typed or typeless access.

The overhauled property only depends on the aforementioned types and
can be tested completely independently of any struct implementation.