introduce proxy property
Remote access of properties is currently realized via the property
proxy class, which actually is a certain type of property instead of a
proxy to a property, as the name implies.

To make this more clear, this commit introduces a proxy property class,
which due to the property changes can be simpler than the property
proxy class.