introduce proxy property iterator
Iterating over properties is currently implemented by the property
iterator class. Like the property itself, it has a circular dependency
with the struct base class. The name also somewhat hides the fact that
it is an iterator for property proxies.

Based on the new proxy property, it is now possible to implement a
corresponding iterator class that only depends on the proxy property
itself. Even though it is very similar to the property iterator, it is
testable without any knowledge of the struct base class implementation.