fix overhauled vector for boolean value type
The overhauled vector does currently not compile for bool value types,
because the std::vector<bool> specialization does not return actual
references to its values. Instead, element access yields a value of a
reference type that abstracts from the underlying bit access. This is
incompatible with the typeless base implementation of the overhauled
vector, because it relies on typeless references being castable to
typed ones, which is not possible with values.

This commit fixes the issue by using uint8_t as the value type when a
bool is used to instantiate the template.

Note that since this introduces an incompatibility to
std::vector<bool>, it should be seen as a temporary solution, until
something better is found.