introduce static descriptor map
Even though the descriptors for static types can be accessed globally,
it is necessary to specify the explicit type to do so. To allow global
access to all static descriptors, this commit introduces a global
descriptor map, which is automatically populated with all static types
that are instantiated (i.e. used) in a given application.