introduce descriptor converter
Until the planned transceiver and connection overhaul, there needs to
be a central preliminary component to provide support for the
conversion of struct descriptor data to descriptors and vice versa.

This commit introduces a descriptor converter type, that is able to do
the necessary conversions on the basis of a specific registry.

Note that the overhauled type system at this point does not yet support
vectors of dynamic struct types.