make overhauled components backwards compatible
To avoid having to migrate all components to the overhauled components
at once, this commit adds several functions to be backwards compatible
to all components that are using the old interfaces.

All functions are marked as deprecated and their usage will generate
corresponding warnings when enabled in the compiler options.