integrate overhauled descriptors into project
This commit replaces all corresponding components with their overhauled
counterparts; completely removing the old ones in the process. To keep
the changes a confined as possible, all dependent components are
integrated into the overhauled descriptor concept by preferring usage
of the backwards compatibility functions.

Note that this also means that the code quality will be worse in many
places until all components are fully migrated to the overhauled
interfaces.

Aside from using the overhauled components, the core integration
changes are the following:

- Code generator templates are adjusted to also generate a descriptor
specialization for their corresponding type
- Code generator config is adjusted to use fundamental type aliases to
be as close to the DOTS modeling langage as possible
- Components for the old vector implementation are being completely
removed
- UUID and chrono descriptors are being replaced by static descriptors
- TStruct and TProperty are being replaced by their static counterparts
- StructProperty is being replaced by PropertyDescriptor
- The transceiver is changed to use a local registry
- Global registry access is being replaced by registry access of the
global transceiver
- Serializers get a struct-only interface while keeping the old ones as
deprecated