remove resize for typeless vectors
Aside from being unnecessarily inefficient, resizing a vector only
works when the value type is default-constructible. If it is not, the
user has to provide a default value that is copied for each created
element. Since this cannot be properly handled at compile time for
typeless vectors, it is better to remove typeless resizing entirely
until there is a good reason to support it.