move dynamic struct impl to dedicated type
The implementation for dynamic struct types is currently part of the
struct descriptor base class. This makes dynamic structs unnecessarily
fiddly and error-prone to use, because they cannot exist without some
form of manual memory management. It also means that they cannot be
part of vectors, because those always require explicit types.

This commits introduces a dynamic struct type that can be used in the
same way as static ones. Because the properties lie in the dynamically
allocated memory for dynamic structs, the property area has to be moved
from the struct base class to the static and dynamic sub-classes. This
also has the advantage that moving a dynamic struct only has the cost
of copying the descriptor and the data pointers.