utilize constexpr metadata in property base class
The property base class is currently accessing the metadata via its
non-constexpr property descriptor. The main drawback of this is that
the offset, and all other metadata attributes, are not available at
compile time, which severely hinders the optimization potential for the
compiler.

This commit utilizes the constexpr metadata in the property base class
as best as possible, leading to a guaranteed compile time evaluation of
the offset for static properties, in addition to all other metadata
attributes, which are also made available for compile time evaluation.

This leads to significantly better optimized code with less function
calls, but also smaller overall code size; in case of the dotsd
reducing the binary size by roughly 20%.