make construct and assign validity check optional
The construct and assign operations for properties are currently
mandatory and sometimes applied redundantly, even when the validity of the
property has already been checked beforehand. Additionally, because
there is a possibility for an exception, calls to these functions are
usually not inlined.

This commit adds a template parameter to the corresponding functions
that allows disabling of the check during compile time, thereby
completely circumventing the code containing the validity check and the
possible exception. These overrides are used in all operations where
the validity is known, either implicitly or explicitly. This is
especially beneficial during initial construction of struct types,
where the properties are guaranteed to be invalid.