add property accessor to generated static structs
Properties are currently accessed by the static struct via offset
arithmetic. This makes it more difficult for the compiler to recognize
which properties are actually accessed, leading to reduced optimization
potential.

This commit moves the property accessor to the generated sub-class,
which can be called in the static struct base class via CRTP. Because
the sub-class knows its properties, they can be returned directly
without having to resort to offset arithmetic at all. The mechanism is
also much simpler, using only a if constexpr via the property type.

In many cases, this change leads to inlining of the entire struct
construction call.