move struct init constructor to generated code
Construction of static struct types via property initializers currently
involves an inherited variadic constructor template that takes a
parameter pack of property initializer, which are unpacked and
perfectly forwarded via a CRTP fold expression call to the property
accessors of the sub class. This is overly complicated and puts a lot
of stress on the compiler and the optimizer, especially with LTO
enabled.

To ease the load on the compiler at least a little bit, it is better to
move the constructor to the generated code, where the properties can be
accessed directly, without the use of CRTP. This reduces the overall
compile times and further increases potential for optimization.