make preload types a transceiver option
Because transceivers should be usable in a variety of contexts, it
might not be desirable to use statically subscribed and published types
in every situation when preloading. Instead, these should be optional
when starting a transceiver. This also allows usage of the preloading
mechanism without statically marked types, e.g. for dynamic structs.

To retain the usability for standard applications, the preloading types
can be created by the same mechanism as before in the application
class.