merge server connection into transceiver
Since the channel has taken over many responsibilities of the server
connection, it does not make much sense anymore to keep it as a
separate class. Because it is tightly bound to the transceiver via the
signals anyway, it can be merged completely into it, thereby avoiding
the signals entirely.