add automatic type management to transceiver
It is currently not tracked anywhere, which types are shared in a
particular moment between two channels. This necessitates that
transceiver users have to make sure that all dependencies of a type are
actually transmitted before transmitting the actual instance.

For the static case, this is currently solved by transmitting all
dependencies in the preloading phase. This is done for all static
types, even if no corresponding instances are actually published at
runtime.

Dynamic components have to keep track themselves which types they have
distributed, leading to redundant implementations.

Since type dependencies are always only required when transmitting an
instance of a particular type, type management can be simplified by
implementing it in the transceiver and incorporating it into the
publishing of instances itself. This also has the benefit of
distributing dependencies lazily only when they are actually used.