avoid storing of server name in connection
The server name is currently passed during construction of a connection
and has to be stored in every instance, despite being the same in each
server application.

Instead, the server name can be passed when starting the async receive
on the connection after it has been created.

Note that this requires to not send the server name in the connect
response to the client, because it is no longer available at this
point. However, since the name was already sent to the client in the
hello message, it is also unnecessary to send it again. Additionally,
the old DOTS implementation checks for this property and only uses it
for logging purposes, which means that the change maintains backwards
compatibility.