avoid storing of client name in channel connection
The client name is currently stored in each channel connection, even
though it is only needed once when sending the initial connect request.

Instead, the name can be passed in the async receive call.

Note that this strictly speaking does not conform to the handshake
definition, since the client is supposed to wait for the hello message
first before sending the first message. However, since both the old and
the overhauled server implementations immediately send the connect
message when starting the async receive, the change is transparent and
thus maintains backwards compatibility.