remove unusable unicast handling
The connection manager currently includes remnants of an old unicast
feature, which allowed one client to send a message directly to another
specific client. However, this feature is not used anywhere and in the
current implementation cannot be used in general. Since the destination
group always has to be valid for previous processing steps to work, the
unicast processing code is unreachable in practice, because the receive
handler always returns earlier, even if a destination client id is set.

Since the feature goes against the publish-subscribe model anyway, it
should be removed entirely and maybe re-added in the protocol overhaul,
if it should turn out to really be required.