deprecate isFromMyself property in DOTS header
The isFromMyself property of the DOTS header (accessed as isOwnUpdate
in events) can currently be used by a local subscriber to determine
whether the instance was sent by itself. However, since this
information depends on the receiving connection, the header has to be
copied, adjusted and then dispatched for every received transmission.

For this reason, the isFromMyself property should be deprecated and
probably removed entirely from the header in the protocol/model
overhaul. Instead, the id comparison result can be passed as an
additional parameter to the dispatcher.