overhaul connection closing and error handling
There is currently no differentiation between expected and unexpected
closing of connections. Additionally, there is no mechanism for a peer
to signal an expected close of its connection.

Instead of only having an error handler, connection users can have a
close handler, where the error is optional for an unexpected close of
the connection.

To signal an expected close to the peer, a connection can send a
DotsMsgError with an error code of 0. Note that this is intended to be
only a temporary solution that is backwards compatible until the
protocol/model overhaul.