differentiate between self and peers in connection
Since connections can be used for clients and servers, the naming of
their names and ids might be confusing. From the point of view of a
client, the id refers to itself, even though it is determined by the
client. For a server, the id refers to the client and not itself.

To better reflect this, this commit introduces a differentiation
between self and peer ids. For a client, its self id is determined by a
client, which in turn is the peer id on the server side.