move dots client global publish out of connection
A connection currently updates the dots client instance according to
its connection state changes. However, because the a connection does
not have access to its connection manager, it has to rely on the a
global publish to update the instance. This is an issue, since
connections are supposed to be usable in non-global contexts.

By using the recently added transition handler, updating the dots
client can now be moved out of the connection to be handled by its
user.