move preloading out of connection
A connection currently handles the preloading during the early
subscribe state. This necessitates that the preloading types have to be
made accessible to the connection, even though they are only relevant
for a client connection at a very specific point. It is also the
reason, why the group handling has to be part of the connection as
well.

Since the preloading is actually just one way of handling the
pre-connect state, it should also rather be the responsibility of the
connection user. This would also allow other types of behavior in
different contexts.

With the recently added transition handler, the preloading handling can
now be delegated to the connection user. This allows the connection to
move closer to its intended design, which is to only be responsible for
establishing the communication between two peers through an arbitrary
channel and not the content of the communication itself.