process instances regardless of number of handlers
The dispatcher currently does not process instances when there is no
corresponding handler present. The idea was to prevent overhead when
there is no subscriber to a particular type.

However, since a subscriber always receives the current container
content upon subscribing, it is counter-intuitive to have a different
behavior for the initial subscriber. A user could reasonably expect
that all dispatched instances are processed and the containers updated,
regardless of how many subscribers there are at any given point.
Additionally, in the current usage scenario only instances that are
relevant to some extent are even dispatched in the first place.