set modified properties on instance creation
The container currently does only set the modified by and modification
time properties in the clone information when the instance is updated,
but not when it is first created. This requires all users to always
check whether the properties are valid.

Since the last operation can always be checked explicitly, it is
reasonable to just use the creating sender and creation time for these
properties.