move dots client and types management
Handling the DOTS types and client instances have an overhead that
might not be desirable for every type of server. Checking for expired
clients in particular is a non-trivial operation that is currently done
once a second along with the connection cleanup.

Since these types are optional and not required for a server in general
to be usable, it should be a feature of the DOTS daemon instead of
every potential server.