fix non-catchable termination for broken UDS pipes
When using UNIX domain sockets, the OS might emit a SIGPIPE to the
application when the pipe is broken. This can happen easily when the
remote endpoint is destroyed unexpectedly. Because the signal is
usually not handled, the application just terminates without any
possibility for recovery. This is especially harmful in servers where
one client can terminate the entire application.

Since there does not seem to be a way to handle the signal for a
specific socket, it has to be disabled globally, in which case the ASIO
socket still signals the broken pipe via an exception that can be
handled normally.