introduce host and guest approach to connections
Applications currently follow the classical server and client approach.
There is usually one server in the form of the dotsd and a number of
client connections to that one server in a starlike topology. However,
since connections can now be established via arbitrary channels, it is
not necessary to restrict them to that concept.

Instead of having a server that has the above mentioned implication in
regards to topology, participants in a DOTS space can be thought of in
terms of hosts and guests. The basic idea is that a host can accept
guest connections, which are all peers within the space, including the
host itself. The only difference between the host and the guests is
that the host can kick guests out if they do not abide by the rules, to
make sure its space stays clean.

In theory, an application does not have to be either a host or a guest.
It could host several spaces at once or be a guest for more than one
host, provided it manages some kind of translation between the spaces.
There could also be different server-like applications, that do
additional local processing in addition to managing its guests, e.g. a
dotsd with tracing. This concept could be extended when protocol is
overhauled in the future, which is currently somewhat restrictive to
fully exploit the possibilities of this approach.

Note that this change is still perfectly capable of providing the
aforementioned starlike server and client applications, since they are
a subset of the host and guest concept. Since they will also probably
still be the most common, the dotsd remains as the standard host in
classical server style.